refactor(constants): switch to index.ts, clean up imports
